refactor(scala): use Scala 2.11 (#3)
For Scala 2.11, 2.11.8 should be used because the patch versions of 2.11
has ordering issues for `circe`.

Note that 2.12 works just as fine too, just that all the `circe` related
versions must be updated to the latest. 2.12 should work just like 2.13,
except the way the macro paradise is enabled in `build.sbt` is
different.

Also 2.11 does not accept trailing commas, and `print` is not available
for `circe` `0.12.0-M3`.