feat(enum): add case where enum relation is inverted
Add a different take on how the enum set-up can be done.

The usual enum is done via `sealed trait` acting as the enum base, but
this has severe limitation on the flexibility of the enum.

Rust can have completely unrelated sets of structs, before gathering
them in a enum base in a completely separate file, and achieve
auto-inferencing on how the serde is done.

Scala cannot easily do this because the enum relation is simulated by
inheritance, which causes an inverted relationship. This inverted
relationship makes it impossible to macro annotate the enum base at
times.