feat(circe): improve the enum variant macro (#14)
Fix by putting always putting fully qualiified inner companion type
rather than using a possible relative qualified one.

Auto deduce fully whether the single param is a case class or not. If
case class, it automatically creates the `apply` method for forwarding
into the inner case class. Otherwise, it does not create.

This eliminates the need to manually specify to include or not.